+++ src/components/load-more.js
@@ -1,0 +1,1 @@
+export const getLoadMoreMarkup = () => `<button class="load-more" type="button">load more</button>`;
